Some formatting improvements
